Create useEditorActions mediator, editing.ts becomes facade
Move all orchestration (openComponent, openPageRoot, openFragment,
save, clear, markDirty, discard, refreshAfterRestore) into a
useEditorActions composable. editing.ts is now a 55-line facade that
re-exports state from the 3 stores and delegates actions.

Key improvements:
- withStashAndRetry extracts the duplicated stash+retry pattern from
  the 3 open methods into one function
- Retry uses bounded exponential backoff (3s, 6s, 12s, max 3 attempts)
  with generation guard instead of infinite setInterval
- Side effects (preview, toast, publishStatus) are explicit in the
  mediator, not buried inside store methods
- All store dependencies flow through one composable instead of 6
  direct store-to-store calls

editing.ts public API unchanged — all 8 consumers work without
modification.
